Adding atomic string implementation

+++ include/atom.h
@@ -5,8 +5,10 @@
 #ifndef ATOM_H_INCLUDED__
 #define ATOM_H_INCLUDED__
 
+#include <stddef.h>
+
 extern const char *atom_string(const char *str);
 
-extern int atom_length(const char *str);
+extern size_t atom_length(const char *str);
 
 #endif /* ATOM_H_INCLUDED__ */
